backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
requests==2.31.0
